refactor: Migrate layout templates to app-level storage
Layout templates are now stored in the app database as user preferences
rather than in the project database. This allows users to create and reuse
page layout templates across all their projects.

Changes:
- Move page_layout_templates and page_layout_elements tables from
  projectSchema to appSchema
- Create new layoutTemplates.ts query file using getAppDatabase()
- Update IPC handlers to ignore projectId parameter (templates are app-wide)
- Remove prep_project_id from PageLayoutTemplate interface, add optional user_id
- Update LayoutDesigner component to not pass prep_project_id when creating templates
- Clean up prep.ts queries file by removing layout template code

Templates are now app-level user preferences that can be shared across all projects
while still being filtered by page_type for relevance.
